Firebase checkPlayerCollision Function
This was configured so that everytime the player moved the postion would 
be checked against the other players in the game. Due to the high number 
of invocations it seems this is not an effiecnt use of Firebase 
functions.

Instead lets look to moving the logic back onto the client and allowing 
the client to deal with the collision ... tbc